backend_discard_obsolete/requirements.txt
# Backend Discard Service Dependencies

# HTTP client for API calls
requests>=2.28.0

# Redis client for efficient queue operations
redis>=5.0.0

# Database connectivity  
supabase>=2.0.0

# Environment configuration
python-dotenv>=1.0.0

# Image processing for report analysis
# Fixed versions for numpy binary compatibility
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0,<11.0.0

# HTML parsing for step-by-step report analysis
beautifulsoup4>=4.12.0